BizPilot_AI_backend/AI_idea_validation/requirements.txt
fastapi 
uvicorn
httpx 
python-dotenv 
pydantic
